Decouple oidc providers, fixes
